Refactor config.rs into a module to split out tests
Split out the ~660 lines of tests from `src/config.rs` into `src/config/tests.rs` to break up the 1500+ line config blob and make it easier to navigate.
